+++ NEWS.md
@@ -3,6 +3,8 @@
 
 * Added new internal function `GetSigActivity`.
 
+* Added new internal function `GetSampleSigActivity`.
+
 # 2.3.1
 * Made the default value for argument `sig.pres.test.nbinom.size` to be NULL in
 various internal and exported functions.